다양한 redux구현하기
