+++ 02_today/src/App.js
@@ -1,5 +1,7 @@
 import { Route, Routes } from "react-router-dom";
 import "./App.css";
+import { CarouselStudy } from "./components/moveComp/CarouselStudy";
+import { ReduxStudy } from "./components/moveComp/ReduxStudy";
 import { DetailPage } from "./pages/DetailPage";
 import { HomePage } from "./pages/HomePage";
 import { Mainpage } from "./pages/Mainpage";
@@ -11,6 +13,8 @@ function App() {
         <Route path="/" element={<Mainpage />} />
         <Route path="/:keyword" element={<HomePage />} />
         <Route path="/detail" element={<DetailPage />} />
+        <Route path="/redux" element={<ReduxStudy />} />
+        <Route path="/carousel" element={<CarouselStudy />} />
       </Routes>
     </div>
   );
